requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arithmetic==0.6.2
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
boto3==1.20.26
botocore==1.23.26
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
greenlet==1.1.2
idna==3.3
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipython-sql==0.4.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.3.3
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-ui-poll==0.2.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
msal==1.16.0
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
powerbiclient==1.0.0
prettytable==0.7.2
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psycopg2==2.9.2
ptyprocess==0.7.0
py4j==0.10.9.2
pycparser==2.21
pydantic==1.8.2
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.0
pyspark==3.2.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-sql==1.3.0
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
s3transfer==0.5.0
Send2Trash==1.8.0
six==1.16.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.7.0
